Define 2d space for chess such as 2d for human and 2d for compute
2d for human is 1a, 1b, 1c, etc.
64 is 8 * 8 = 64 8+8+8=24+abc=3=idx=27, whatever. so you can probably
   loop only 8 steps with plus location with abcdefgh 2d for compute on
the other hand is on the fly chess piece available capture of cell or
not is block probably

Define and create functions for modifying peice moves
yeah so i have change from using dynamic to static memory
i guess you could say that you just have to manipulate static
values to make the chess piece move from here to there.

create a func that modifies coordinate xy[2] based on the index
use this for moving from place to places.

Fix: change the names and test them in another file

todo: make use of getchar function to get input from the
user and then convert and reduce the input to be usable.
for example:
input: e61b means  sixth pawn goes to index 57,
	n25b means 2nd knight goes to index 41

todo: come up with a way to create the game loop.