Create more auxillary functions that helps with later functions
this eventually creates the application.
todo:
create a function that draws the chess so you can see the output
later on when manipulating states. this could help with testing
but I ain't gonna do some testing, just testing values only with
stdio would be enough really? I could have preferred it but yeah.

create draw chess function that draws the chess in ascii